Split SpotsDelegates into multiple files
