+++ Sources/Shared/Protocols/SpotsFocusDelegate.swift
@@ -1,0 +1,4 @@
+public protocol SpotsFocusDelegate: class {
+  var focusedSpot: Spotable? { get set }
+  var focusedItemIndex: Int? { get set }
+}
